refatorando e adicionando novos end points

+++ src/main/java/br/com/beautystyle/agendamento/controller/ClientController.java
@@ -1,8 +1,11 @@
 package br.com.beautystyle.agendamento.controller;
 
+import br.com.beautystyle.agendamento.controller.dto.ClientDto;
 import br.com.beautystyle.agendamento.controller.form.ClientForm;
-import br.com.beautystyle.agendamento.model.Client;
+import br.com.beautystyle.agendamento.model.entity.Client;
 import br.com.beautystyle.agendamento.repository.ClientRepository;
+import br.com.beautystyle.agendamento.repository.CompanyRepository;
+import br.com.beautystyle.agendamento.repository.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,48 +25,47 @@ public class ClientController {
     @Autowired
     private ClientRepository clientRepository;
 
-    @GetMapping
-    public List<Client> findAll() {
-        return  clientRepository.findAll();
-    }
+    @Autowired
+    private EventRepository eventRepository;
+
+    @Autowired
+    private CompanyRepository companyRepository;
+
 
     @GetMapping("/{id}")
-    public ResponseEntity<Client> findById(@PathVariable Long id) {
-        Optional<Client> clientById = clientRepository.findById(id);
-        return clientById
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+    public List<ClientDto> findByCompanyId(@PathVariable Long id) {
+        List<Client> clientList = clientRepository.findByCompanyId(id);
+        return  ClientDto.convert(clientList);
     }
 
     @PostMapping
     @Transactional
-    public ResponseEntity<Client> insert(@RequestBody @Valid ClientForm clientForm, UriComponentsBuilder uriBuilder) {
-        Client savedClient = clientRepository.save(clientForm.converter());
+    public ResponseEntity<ClientDto> insert(@RequestBody @Valid ClientForm clientForm, UriComponentsBuilder uriBuilder) {
+        Client savedClient = clientRepository.save(new Client(clientForm));
         URI uri = uriBuilder.path("/client/{id}")
                 .buildAndExpand(savedClient.getClientId())
                 .toUri();
-        return ResponseEntity.created(uri).body(savedClient);
+        return ResponseEntity.created(uri).body(new ClientDto(savedClient));
     }
 
-    @PostMapping("/salva_lista")
+    @PostMapping("/client_list")
     @Transactional
-    public ResponseEntity<List<Client>> insertAll(@RequestBody @Valid List<ClientForm> clientList, UriComponentsBuilder uriBuilder) {
-        List<Client> newClients = new ArrayList<>();
-        clientList.forEach(client -> newClients.add(client.converter()));
-        List<Client> savedClientList = clientRepository.saveAll(newClients);
+    public ResponseEntity<List<ClientDto>> insertAll(@RequestBody @Valid List<ClientForm> clientList, UriComponentsBuilder uriBuilder) {
+        List<Client> newClients = Client.convert(clientList);
+        List<Client> savedClients = clientRepository.saveAll(newClients);
         URI uri = uriBuilder.path("/client/{id}")
-                .buildAndExpand(savedClientList)
+                .buildAndExpand(savedClients)
                 .toUri();
-        return ResponseEntity.created(uri).body(savedClientList);
+        return ResponseEntity.created(uri).body(ClientDto.convert(savedClients));
     }
 
     @PutMapping
     @Transactional
-    public ResponseEntity<Client> update(@RequestBody @Valid Client client) {
-        Optional<Client> clientById = clientRepository.findById(client.getClientId());
+    public ResponseEntity<ClientDto> update(@RequestBody @Valid ClientDto clientDto) {
+        Optional<Client> clientById = clientRepository.findById(clientDto.getApiId());
         if (clientById.isPresent()) {
-            Client update = client.update(clientRepository);
-            return ResponseEntity.ok(update);
+            Client client = clientDto.update(clientRepository);
+            return ResponseEntity.ok(new ClientDto(client));
         }
         return ResponseEntity.notFound().build();
     }
@@ -73,7 +75,13 @@ public class ClientController {
     public ResponseEntity<?> delete(@PathVariable Long id) {
         Optional<Client> clientById = clientRepository.findById(id);
         if (clientById.isPresent()) {
-            clientRepository.deleteById(id);
+            Client client = clientById.get();
+            if(client.getUserId()!=null){
+               client.removeAssociation(eventRepository);
+                clientRepository.deleteById(id);
+            }else{
+                return ResponseEntity.notFound().build();
+            }
             return ResponseEntity.ok().build();
         }
         return ResponseEntity.notFound().build();
